MISC: Rename example runscript names
